move mergeNotes/replaceVoice to sequences
